sysbuild: Functionality to pass variables from sysbuild directly to projects cmake
Small change that gives a possibility to create list of variables
that should be passed directly to projects cmakes.
Such variables would be available before Zephyr project inclusion
inside the project itself.
The list needs to have a form <namespace>_USER_SHARED_VARS and needs
to be defined before ExternalZephyrProject_Add.
When it is used to pass variables to main application it needs to defined
before Sysbuild package (in sysbuild/CmakeLists.txt).

Signed-off-by: Radosław Koppel <r.koppel@k-el.com>